Conversion of software specific modified peptide entries

Modified peptide entries are converted to a common text representation

convert_modified_peptides(
  input_df,
  software = c("MaxQuant", "PD", "DIA-NN", "Spectronaut")
)

Arguments

input_df

A tibble with modified peptide level information. For MaxQuant: evidence.txt, for PD: PSMs.txt with R-friendly headers enabled, for DIA-NN and Spectronaut default output reports.

software

The used analysis software for the input_df - MaxQuant, PD, DIA-NN or Spectronaut. Default is MaxQuant.

Value

This function returns the original submitted tibble - input_df - including two new columns:

traceR_mod.peptides - software-independent standardized text for modified peptide entries.

traceR_mod.peptides_unknownMods - logical value, if TRUE: a modification is detected, which is not converted to a standardized text.

Details

The input entries are converted to a software independent format. The generated entries are appended to the submitted dataframe. Conversion of modifications is currently only available for UniMod:35 and UniMod:4. Other modifications will not be converted to standardized format.

Author

Oliver Kardell

Examples

# Load libraries
library(dplyr)
library(stringr)
library(tidyr)
library(tibble)

# MaxQuant example data
data <- tibble::tibble(
 "Modified sequence" = c("_AACLLPK_",
  "_ALTDM(Oxidation (M))PQM(Oxidation (M))R_",
  "ALTDM(Dummy_Modification)PQMK"),
  Charge = c(2,2,3)
)

# Conversion
convert_modified_peptides(
 input_df = data,
 software = "MaxQuant"
)
#> # A tibble: 3 x 5
#>   `Modified sequence` Precursor.Id    Charge traceR_mod.pepti~ traceR_mod.pepti~
#>   <chr>               <chr>            <dbl> <chr>             <lgl>            
#> 1 _AACLLPK_           AACLLPK2             2 AACLLPK           FALSE            
#> 2 _ALTDM(Oxidation (~ ALTDM(Oxidatio~      2 ALTDM(UniMod:35)~ FALSE            
#> 3 ALTDM(Dummy_Modifi~ ALTDM(DummyMod~      3 ALTDM(DummyModif~ TRUE