Conversion of software specific proteinGroups

ProteinGroups are converted to a common text representation

convert_proteingroups(
  input_df,
  software = c("MaxQuant", "DIA-NN", "Spectronaut", "PD")
)

Arguments

input_df

A tibble with proteinGroup level information. For MaxQuant: proteinGroups.txt, for PD: PSMs.txt with R-friendly headers enabled, for DIA-NN and Spectronaut default output reports.

software

The used analysis software for the input_df - MaxQuant, PD, DIA-NN or Spectronaut. Default is MaxQuant.

Value

This function returns the original submitted tibble - input_df - including one new column:

traceR_proteinGroups - software-independent standardized text for proteinGroups.

Details

The input entries are converted to a software independent format. The generated entries are appended to the submitted dataframe.

Author

Oliver Kardell

Examples

# Load libraries
library(dplyr)
library(stringr)
library(comprehenr)
library(tibble)

# MaxQuant example data
data <- tibble::tibble(
"Protein IDs" = c("A0A075B6P5;P01615;A0A087WW87;P01614;A0A075B6S6", "P02671", "P02672"),
id = c(26, 86, 17)
)

# Conversion
convert_proteingroups(
 input_df = data,
 software = "MaxQuant"
)
#> # A tibble: 3 x 3
#>   `Protein IDs`                          id traceR_proteinGroups                
#>   <chr>                               <dbl> <chr>                               
#> 1 A0A075B6P5;P01615;A0A087WW87;P0161~    26 A0A075B6P5;A0A075B6S6;A0A087WW87;P0~
#> 2 P02671                                 86 P02671                              
#> 3 P02672                                 17 P02672